requirements.txt
Flask
flask-cors
flask-restful
langchain
transformers
faiss-cpu
sentence-transformers
gunicorn
